refactor: merged the card component tests into one file
