+++ src/_test_/card.test.tsx
@@ -3,7 +3,30 @@ import { describe, expect, it } from 'vitest';
 import { Card } from '../components/card/card';
 import type { CardProps } from '../types/types';
 
-describe('Card handles missing or undefined data gracefully', () => {
+describe('Test card component', () => {
+  // Displays all card fields correctly
+  it('Correctly displays item names and descriptions', () => {
+    const mockData = {
+      name: 'Test Animal',
+      uid: '12345',
+      avian: true,
+      canine: false,
+      feline: true,
+      earthAnimal: false,
+      earthInsect: true,
+    };
+    render(<Card {...mockData} />);
+    expect(screen.getByText('Test Animal')).toBeInTheDocument();
+    expect(screen.getByText(/UID:/i)).toBeInTheDocument();
+    expect(screen.getByText('12345')).toBeInTheDocument();
+    expect(screen.getByText(/Avian:/i).textContent).toMatch(/true/i);
+    expect(screen.getByText(/Canine:/i).textContent).toMatch(/false/i);
+    expect(screen.getByText(/Feline:/i).textContent).toMatch(/true/i);
+    expect(screen.getByText(/Earth Animal:/i).textContent).toMatch(/false/i);
+    expect(screen.getByText(/Earth Insect:/i).textContent).toMatch(/true/i);
+  });
+
+  // Renders default values when props are missing
   it('renders fallback values for missing properties', () => {
     const partialData: Partial<CardProps> = {};
